refactor(docs-gen): split extractor modules
